Rename logo color file
