+++ server/services/icons/FeatherIconsService.ts
@@ -1,6 +1,6 @@
 import feather, { FeatherIcon } from 'feather-icons';
 import IconsService from './IconsService';
-import { normalizeColor } from '../setLogoColor';
+import { normalizeColor } from '../logoColor';
 
 class FeatherIconsService extends IconsService {
   public static async getIcon(slug: string, color: string|null = null):
