+++ server/services/logoColor.ts
